requirements.txt
Django==5.1.4
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
django-liveview==2.1.5
redis==5.2.1
